Size_t is wrong.
git-svn-id: https://subversion.assembla.com/svn/qmcdev/trunk@2089 e5b18d87-469d-4833-9cc0-8cdfa06e9491

+++ src/Utilities/PooledData.h
@@ -25,7 +25,7 @@ using namespace std;
 template<class T>
 class PooledData: public std::vector<T> {
 
-  size_t Current;
+  int Current;
 
 public:
